+++ src/ember/grid.py
@@ -1656,7 +1656,7 @@ class Grid(_LabelledList):
                     t=block.T_nd,
                     mu=block.mu_nd,
                     cp=block.cp_nd,
-                    pr_lam=block.fluid._Pr,
+                    kappa=block.kappa_nd,
                     pr_turb=1.0,
                     xlength=block.xlen_sq_nd,
                     vol=block.vol_nd,